extract the council visualisation
